requirements.txt
ccxt==4.1.78
pandas>=2.0.0
numpy>=1.24.0
python-telegram-bot==20.7
flask==2.0.1
gunicorn==20.1.0
Werkzeug==2.0.3